+++ crawler/pom.xml
@@ -13,6 +13,7 @@
     <version>0.0.1-SNAPSHOT</version>
     <name>crawler</name>
     <description>Crawler for ycombinator data</description>
+
     <properties>
         <java.version>11</java.version>
     </properties>
@@ -34,7 +35,6 @@
             <artifactId>jsoup</artifactId>
             <version>1.14.3</version>
         </dependency>
-
     </dependencies>
 
     <build>
